+++ pkg/iocounter/counter.go
@@ -1,3 +1,4 @@
+// Package iocounter provides an io.Writer that tracks how many bytes have been written to it.
 package iocounter
 
 // Counter counts a number of bytes during an IO operation.
