+++ pkg/iocounter/writer.go
